Add privacy policy text
