+++ modules/features/account/src/main/java/au/com/shiftyjelly/pocketcasts/account/onboarding/upgrade/OnboardingUpgradeHelper.kt
@@ -34,18 +34,24 @@ import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.constraintlayout.compose.ConstrainedLayoutReference
 import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.constraintlayout.compose.ConstraintLayoutScope
 import au.com.shiftyjelly.pocketcasts.account.R
+import au.com.shiftyjelly.pocketcasts.compose.components.Clickable
+import au.com.shiftyjelly.pocketcasts.compose.components.ClickableTextHelper
 import au.com.shiftyjelly.pocketcasts.compose.components.TextH30
 import au.com.shiftyjelly.pocketcasts.compose.components.TextP60
 import au.com.shiftyjelly.pocketcasts.compose.extensions.brush
+import au.com.shiftyjelly.pocketcasts.preferences.Settings
 import au.com.shiftyjelly.pocketcasts.images.R as IR
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
@@ -420,5 +426,43 @@ object OnboardingUpgradeHelper {
                 tint = Color.White,
             )
         }
+    }
+
+    @Composable
+    fun PrivacyPolicy(
+        color: Color,
+        textAlign: TextAlign,
+        modifier: Modifier = Modifier,
+        lineHeight: TextUnit = 16.sp,
+    ) {
+        val privacyPolicyText = stringResource(LR.string.onboarding_plus_privacy_policy)
+        val termsAndConditionsText = stringResource(LR.string.onboarding_plus_terms_and_conditions)
+        val text = stringResource(
+            LR.string.onboarding_plus_continuing_agrees_to,
+            privacyPolicyText,
+            termsAndConditionsText
+        )
+        val uriHandler = LocalUriHandler.current
+        ClickableTextHelper(
+            text = text,
+            color = color,
+            lineHeight = lineHeight,
+            textAlign = textAlign,
+            clickables = listOf(
+                Clickable(
+                    text = privacyPolicyText,
+                    onClick = {
+                        uriHandler.openUri(Settings.INFO_PRIVACY_URL)
+                    }
+                ),
+                Clickable(
+                    text = termsAndConditionsText,
+                    onClick = {
+                        uriHandler.openUri(Settings.INFO_TOS_URL)
+                    }
+                ),
+            ),
+            modifier = modifier,
+        )
     }
 }
